requirements.txt
gensim==4.3.2
kagglehub==0.3.10
matplotlib==3.8.4
neo4j==5.26.0
nltk==3.9.1
nltk==3.8.1
numpy==2.2.4
pandas==1.5.3
Pillow==11.1.0
scikit_learn==1.4.2
scikit_surprise==1.1.4
scipy==1.15.2
seaborn==0.11.2
surprise==0.1
torch==2.6.0
torch==2.2.2+cpu
tqdm==4.67.1
tqdm==4.66.6
tqdm==4.66.5
tqdm==4.66.4
wordcloud==1.9.4
